Finished 5 and 6: place markers when correct character is selected

+++ javascript_jquery/wheres_waldo/db/migrate/20170710153236_create_characters.rb
@@ -2,8 +2,8 @@ class CreateCharacters < ActiveRecord::Migration[5.1]
   def change
     create_table :characters do |t|
       t.string :name
-      t.decimal :x_position
-      t.decimal :y_position
+      t.integer :x_position
+      t.integer :y_position
       t.integer :puzzle_id
 
       t.timestamps
